修复编译警告

+++ src/Http/HttpSession.cpp
@@ -515,7 +515,7 @@ void HttpSession::sendResponse(int code,
     if(no_content_length){
         //http-flv直播是Keep-Alive类型
         bClose = false;
-    }else if(size >= SIZE_MAX || size < 0 ){
+    }else if((size_t) size >= SIZE_MAX || size < 0 ){
         //不固定长度的body，那么发送完body后应该关闭socket，以便浏览器做下载完毕的判断
         bClose = true;
     }
